requirements.txt
certifi==2018.4.16
oogle-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
httplib2==0.11.3
auth2client==4.1.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
python-dateutil==2.6.0
requests==2.19.1
urllib3==1.23
